Enhance DGPS Age Handling in GNSSProcessor
- Changed the type of ageDGPS from uint16_t to float for improved precision.
- Updated the calculation of DGPS age to use floating-point values in various message parsers.
- Introduced logging for maximum and average DGPS age, along with message count for better monitoring.
- Reset maximum DGPS age and message count after logging to maintain accurate statistics.

+++ lib/aio_navigation/GNSSProcessor.h
@@ -44,7 +44,7 @@ public:
         uint8_t fixQuality; // 0=invalid, 1=GPS, 2=DGPS, 4=RTK, 5=Float
         uint8_t numSatellites;
         float hdop;
-        uint16_t ageDGPS; // seconds since last DGPS update
+        float ageDGPS; // seconds since last DGPS update
 
         // Velocity data
         float speedKnots;
